refactor: continue; todo: motors controller PWM class

+++ esp32_drone/logger.cpp
@@ -2,7 +2,7 @@
 
 Logger logger;
 
-void Logger::begin(unsigned long baudRate)
+void Logger::begin(uint32_t baudRate)
 {
     Serial.begin(baudRate);
     while (!Serial)
